requirements.txt
Flask==2.3.3
python-nmap==0.7.1
requests==2.31.0
Flask-WTF==1.1.1
python-dotenv==1.0.0
Werkzeug==2.3.7
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
urllib3==2.0.4
